Add `pricingPlanId` to every relevant course

+++ src/collections/courses.ts
@@ -52,8 +52,8 @@ export default defineCollection({
           .array(reference("authors"))
           .min(1, "At least 1 instructor is required")
           .default(["daniel-ciocirlan"]),
-        price: z.number().positive().optional(),
-        pricingPlanId: z.number().positive().optional(),
+        price: z.number().int().positive().optional(),
+        pricingPlanId: z.number().int().positive().optional(),
         purchaseLink: z.string().url(),
         question: z
           .object({
@@ -62,7 +62,6 @@ export default defineCollection({
           })
           .strict()
           .optional(),
-        // teachable_pricing_plan_id: z.number().int(),
         title: z.string(),
       })
       .strict(),
